Add kstream-take procedure and unit tests for it
*kstream-take* receives as parameter a *kstream* and a integer *n* to
take elements from this stream. The final result is a conventional
racket list constructed by reading the stream *n* times.

If a stream is finite and has only *k* elements, the procedure returns
a list of *min(n, k)* elements.

A set of unit tests covering the basic usage is provided.

Usage:

```racket
> (require "prime.rkt" "numbers.rkt" "stream.rkt")
> (kstream-take (naturals) 10)
'(1 2 3 4 5 6 7 8 9 10)
> (kstream-take (primes-stream) 10)
'(2 3 5 7 11 13 17 19 23 29)
```